research/band_diagram_scan/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bandlib"
version = "0.1.0"
description = "2D photonic band-diagram scan tooling built around MPB/Meep"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "RnLe" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "numpy>=1.24",
    "h5py>=3.8",
    "PyYAML>=6.0",
    "typer>=0.9",
    "tqdm>=4.65",
    "rich>=13.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.scripts]
bandlib = "bandlib.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["bandlib*"]
